Beacon range shrinks.
The way neighbors are determined is by getting every beacon in range of the ant and then checking that the  ant is in the range of the beacon itself. No checks are done with the range of currBeacon (this wouldn't ensure the correct propagation of pheromones as what is  now currBeacon might be out of the neighborhood of the next beacon visited stopping the propagation).