Model parameters changed to improve performance.
Now the system is able to sustain an optimized path between the nest and food source. The increase of pExplore improved the ability of connecting two separate Beacon networks to restore the path between stuck ants and the food source or nest.

As a new metric is added an avg travel length that weights more recent data ( is less affected by the high initial values needed to generate a path between food and nest)